ref: Split controller and server into two actors
