+++ server/src/actors/mod.rs
@@ -2,4 +2,5 @@ pub mod controller;
 pub mod events;
 pub mod keys;
 pub mod project;
+pub mod server;
 pub mod upstream;
